Move gmail to common folder
